remove unneccessary audits. keep their tests.
