+++ lighthouse-core/config/default.json
@@ -71,15 +71,13 @@
     "user-timings",
     "screenshots",
     "critical-request-chains",
+    "webapp-install-banner",
     "manifest-exists",
     "manifest-background-color",
     "manifest-theme-color",
     "manifest-icons-min-192",
-    "manifest-icons-min-144",
     "manifest-name",
-    "manifest-short-name",
     "manifest-short-name-length",
-    "manifest-start-url",
     "theme-color-meta",
     "content-width",
     "deprecations",
@@ -208,26 +206,8 @@
       }
     }, {
       "name": "User can be prompted to Add to Homescreen",
-      "description": "While users can manually add your site to their homescreen in the browser menu, the [prompt (aka app install banner)](https://developers.google.com/web/updates/2015/03/increasing-engagement-with-app-install-banners-in-chrome-for-android) will proactively prompt the user to install the app if the below requirements are met and the user has visited your site at least twice (with at least five minutes between visits).",
-      "see": "https://github.com/GoogleChrome/lighthouse/issues/23",
       "audits": {
-        "service-worker": {
-          "expectedValue": true,
-          "weight": 1
-        },
-        "manifest-exists": {
-          "expectedValue": true,
-          "weight": 1
-        },
-        "manifest-start-url": {
-          "expectedValue": true,
-          "weight": 1
-        },
-        "manifest-icons-min-144": {
-          "expectedValue": true,
-          "weight": 1
-        },
-        "manifest-short-name": {
+        "webapp-install-banner": {
           "expectedValue": true,
           "weight": 1
         }
